Split out html-events crate
