+++ packages/html-events/src/events/mounted.rs
@@ -207,13 +207,7 @@ impl MountedData {
     }
 }
 
-use dioxus_core::Event;
-
 use crate::geometry::{PixelsRect, PixelsSize, PixelsVector2D};
-
-pub type MountedEvent = Event<MountedData>;
-
-pub use super::onmounted as onmount;
 
 /// The MountedResult type for the MountedData
 pub type MountedResult<T> = Result<T, MountedError>;
