+++ packages/html-events/src/events/pointer.rs
@@ -1,10 +1,9 @@
-use dioxus_core::Event;
 use keyboard_types::Modifiers;
 
-use crate::{geometry::*, input_data::*, *};
-
-/// A synthetic event that wraps a web-style [`PointerEvent`](https://developer.mozilla.org/en-US/docs/Web/API/PointerEvent)
-pub type PointerEvent = Event<PointerData>;
+use crate::{
+    geometry::*, input_data::*, InteractionElementOffset, InteractionLocation,
+    ModifiersInteraction, PointerInteraction,
+};
 
 pub struct PointerData {
     inner: Box<dyn HasPointerData>,
