+++ packages/html-events/src/events/touch.rs
@@ -1,10 +1,8 @@
-use dioxus_core::Event;
 use keyboard_types::Modifiers;
 
 use crate::geometry::*;
 use crate::{InteractionLocation, ModifiersInteraction};
 
-pub type TouchEvent = Event<TouchData>;
 pub struct TouchData {
     inner: Box<dyn HasTouchData>,
 }
