+++ packages/html-events/src/geometry.rs
